Protobuf dataformat native support #789
