+++ docs/modules/ROOT/partials/reference/dataformats/protobuf.adoc
@@ -2,11 +2,11 @@
 // This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 :cq-artifact-id: camel-quarkus-protobuf
 :cq-artifact-id-base: protobuf
-:cq-native-supported: false
-:cq-status: Preview
+:cq-native-supported: true
+:cq-status: Stable
 :cq-deprecated: false
 :cq-jvm-since: 1.0.0
-:cq-native-since: 1.0.0
+:cq-native-since: 1.5.0
 :cq-camel-part-name: protobuf
 :cq-camel-part-title: Protobuf
 :cq-camel-part-description: Serialize and deserialize Java objects using Google's Protocol buffers.
