+++ integration-tests/protobuf/src/main/proto/addressbook.proto
